Have made Dave's change requests as well as tweaked the Demo.
