+++ AudioKit/Common/Internals/Utilities/AVAudioTimeShim.swift
@@ -82,6 +82,22 @@ extension AVAudioTime {
         }
         return nil
     }
+    
+    
+    /// convert an AVAudioTime object to seconds with a hostTime reference
+    open func toSeconds(hostTime: UInt64) -> Double {
+        return AVAudioTime.seconds(forHostTime: self.hostTime - hostTime)
+    }
+    
+    // convert seconds to AVAudioTime with a hostTime reference
+    open class func secondsToAudioTime(hostTime: UInt64, time: Double) -> AVAudioTime {
+        // Find the conversion factor from host ticks to seconds
+        var timebaseInfo = mach_timebase_info()
+        mach_timebase_info(&timebaseInfo)
+        let hostTimeToSecFactor = Double(timebaseInfo.numer) / Double(timebaseInfo.denom) / Double(NSEC_PER_SEC)
+        let out = AVAudioTime(hostTime: hostTime + UInt64(time / hostTimeToSecFactor))
+        return out
+    }
 
 }
